Slack probe notifications
Each probe now has an option to choose one of 'Email' or 'Slack' for
'Notify the author' and 'Notify others'.

Sundeck is providing one external function and a Slack app that can be
installed if you create a Sundeck account.

'Notify the author' for Slack depends on the email address you have for
Snowflake matching the email address on your Slack profile in the
workspace you've installed the Sundeck app to. 'Notify others' can
accept Slack user's email addresses, slack usernames, and channel names.

The INTERNAL.PROBES table has to be migrated to handle the "mechanism"
for each method of notification. A procedure will perform that migration
in an idempotent fashion, defaulting all existing probe notifications to
be 'Email' notifications.

Closes #25, #74