Correct the reversed values for RECORD_TYPE (#228)
Creates new "enum" values for RECORD_TYPE and uses the old values as the
hint to correct the user-facing views without rewriting all of the data.

Required unsetting the (implied) `VARCHAR(8)` on the materialized tables
for the RECORD_TYPE column. The natural table migration logic does not
catch this case and it felt unnecessarily complex to try to update that
for this one-off.