Ensure ENABLED is set for labels (#289)
This column has been there since day 1, but it looks like we forgot to
set it. We don't support disabling labels (but I could see us doing so
in the future). Set all labels as being "enabled" for now.

If it causes user confusion, we can drop it from the CATALOG.LABELS in a
follow-on.

Closes #286