Fix mis-ordered columns between internal views and staging table (#265)
Force a stable ordering of new columns for view->table migrations.

We had lots of iteration due to bugs since v1.19, but those bugs prevented any new columns on the Snowflake views from being seen inside of OpsCenter. This has the benefit of letting us avoid solving any tricky migration problems because we know it was impossible to have any new columns that were added to the materialized views in the wrong order.