Ensure that the prettified autoscaling values don't get interpreted as autoscaling being active
The UI tries to avoid showing scale min/max equal to 0 when autoscaling
is not enabled; however, this cause the value of 1 (not zero) to be
persisted on update. This causes the scheduling task to think that it
has to include these values on `alter warehouse` which fails (since the
warehouse is not enabled for autoscaling).

Push all of this down onto the WarehouseSchedules instance and add some
unit tests.

Closes #392