Move the warehouse schedules table creation out of python
We moved the table creation into finalize_setup() because we cannot call
a python procedure during the native app setup script. However, this
created a different problem that this procedure needs to be called (e.g.
the UI opened) before the user is able to see/use the catalog views.

Lift the table/view creation out of python and put it into SQL again.