Include ENABLED for catalog.labels (#488)
We have two create view statements:
* This one that is executed during install
* The label migration invoked during admin.finalize_setup()

These two were inconsistent: one excluded the ENABLED column and the
other did not.

I previously had made the call that we should continue to include the
ENABLED flag even though we don't currently use it. So, we will make
sure that the ENABLED flag is present in both create view statements.