Move procedures migrating non-versioned tables outside of finalize_setup (#499)
On install/upgrade, we want to rely on the setup script to migrate the
schema for all internal tables that are NOT relying on the snowflake
database.

Currently, these migration procedures are only called when
finalize_setup() runs. This is problematic as this only runs when user
invokes it or a user opens the native app. Calling these procedures in
the setup script itself (not finalize_setup()) will ensure no disconnect
between the frontend and the "backend" (SQL procedures + Python code).