Fix the procedure to properly run in the setup script.
Still not sure why this only caused problems in the setup script (and
not on its own). Must be something tricky with the manner in which setup
script is called.

I also moved calling this procedure into 004_labels (instead of at the
end in 091_migrations) because I think this will prevent the need for us
to care later if we still have semi-broken labels.