Fix any duplicate labels that were caused before #496 (#503)
* Fix any duplicate labels that were caused before #496

My fix only prevented future duplicates from being created but didn't
address any duplicate labels already in the internal.labels table.

* Comment the script

* Cleanup

* Fix the procedure to properly run in the setup script.

Still not sure why this only caused problems in the setup script (and
not on its own). Must be something tricky with the manner in which setup
script is called.

I also moved calling this procedure into 004_labels (instead of at the
end in 091_migrations) because I think this will prevent the need for us
to care later if we still have semi-broken labels.

* Avoid the alter table

* add unique id

* Fix unit test

* fix tests again

* one more time

* fix last test

---------

Co-authored-by: Ryan Murray <rymurr@gmail.com>