Correct issue where Snowflake skips warehouse suspend completed event message
Our expectation of the event model for warehouse events history was:

1. wh suspend start
1. wh suspend complete
1. wh resume start
1. wh resume complete
1. wh suspend start
1. wh suspend complete
1. etc

It looks like there is a pattern where this can be the sequence of events:

1. wh suspend start
1. wh resume start
1. wh resume complete

This patterns skips the “suspend complete” before firing a “resume start”. When analyzing internal data, this seems to happen more rarely in the past few months suggesting that it might have been a bug in Snowflake that was fixed. 

This issue only impacts warehouse events, not cluster events (which don’t have a start and completed, only a completed). However, it impacts both our tables because we contextualize cluster sessions within warehouse sessions in our first match recognize to facilitate the generation of a warehouse session id to be used for later joins between warehouse sessions and cluster sessions.

Proposing a solution to  use “suspend started” as the end of a warehouse session instead of “suspend completed”