Always re-create external functions during UPGRADE_CHECK()
The linking to Sundeck path will set a new definition for
internal.get_ef_token() but we don't want to proactively re-create
the external functions because this takes a few seconds.

Instead, recreate the functions in UPGRADE_CHECK rather than in
FINALIZE_SETUP which UPGRADE_CHECK may or may-not call.