Change SIMPLE_DATA_EVENTS_MAINTENANCE to run every 12hrs (#656)
* Change SIMPLE_DATA_EVENTS_MAINTENANCE to run every 12hrs

Also, move the refreshing of internal.sfwarehouses into
WAREHOUSE_LOAD_MAINTENANCE (since that task reads those warehouses).

* Update the SDEM task rows in mat_status to reflect 12hr task schedule

Leave a TODO to remove an unused procedure after the next public release;
else there is a chance the old tasks are still referring to this
procedure. This procedure is no longer used after the task_log was
introduced.