Migrate all NTZ in task_log back to LTZ (#660)
The times in admin.materialization_status don't have timezones but are
set to the current session timezone. This leads to confusion where
downstream systems may interpret the time as UTC instead of the actual
timezone.

To avoid weirdness of TIMESTAMP_LTZ values changes through a procedure
call, this avoids passing start_time back when in FINISH_TASK.

There is no way to correct the data in-place. We have to do this before
we release the native app publicly.